pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alphapil"
version = "0.1.0"
description = "An asynchronous, template-based image generation engine"
authors = [{name = "AlphaPIL Team", email = "team@alphapil.dev"}]
license = "MIT"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "Pillow>=9.0.0",
    "aiohttp>=3.8.0",
    "discord.py>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "isort>=5.10.0",
]

[project.urls]
Homepage = "https://github.com/MeraLabs/AlphaPIL"
Repository = "https://github.com/MeraLabs/AlphaPIL"
Documentation = "https://alphapil.readthedocs.io"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88
